feature: 회사 검색 및 리스트 - Home에서만 가능하도록
